[build; WIP] work towards fixing mpv build; [events] temporarily feature-gate events; [wrapper] remove convenience function enums in favor of individual methods; [meta] replace try! with ?, update
The main problem with the build script was that, for whichever reason, it cleared the environment which resulted in having no package etc. locations. Another problem when compiling statically was that the environment variable TARGET was set by cargo to something the build script interpreted incorrectly.

The feature-gate is temporary, in that once there's the constant dependent type-system, a constant bool will be used, thus having all the advantages of the feature plus a decision of whether to use events per mpv instance.

Single methods are easier to use, document, and are more idiomatic. In the process of this, a few special case structs were also removed.