[events] Fix UB when casting from raw pointer
Works towards resolving #2, however does not seem to solve all cases.
When an Event with e.g. the id 9 was returned by mpv_wait_event, this
caused UB such as taking the QueueOverflow branch. This also reverts
a previous commit, that removed the disabling of events inaapropriatly,
as the events do have to be disabled explicitly.